Eval integer literal
[日記]
- evaluator に来た
- Go で int64 だったので i64 に合わせた
- Go だと eval(ast.Node) なんだけど、ast::Node を作成していないのでできない
  - し、trait Node してもそこから具象クラスを取れないのでダメ
- trait Eval を作って、Node として渡ってくるものに eval() メソッドを生やすようにした

+++ src/ast.rs
@@ -136,7 +136,7 @@ impl fmt::Display for Identifier {
 
 #[derive(Debug, PartialEq)]
 pub struct IntegerLiteral {
-    pub value: isize,
+    pub value: i64,
 }
 
 impl fmt::Display for IntegerLiteral {
